fix existence check for keys (stat)
In preparation for #2.

A stat request (and response is also .present?) if the file that is
stat'ed is not, at least in my case.

Probably one wants to know whether the stat request succeeded, the
response is allright and then if the response indicates that the "key"
is present and a file; but the implemented check seems at least to be
better than the previous implementation.